rename content_type.py -> generic.py
