+++ dynamic_rest/fields/__init__.py
@@ -1,2 +1,2 @@
 from dynamic_rest.fields.fields import *  # noqa
-from dynamic_rest.fields.content_type import *  # noqa
+from dynamic_rest.fields.generic import *  # noqa
